move the final device_auth thing to nexus/src/app
