--- nexus/src/external_api/device_auth.rs
@@ -1,110 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-//! Entrypoints for the OAuth 2.0 Device Authorization Grant flow.
-//!
-//! These are endpoints used by the API client per se (e.g., the CLI),
-//! *not* the user of that client (e.g., an Oxide rack operator). They
-//! are for requesting access tokens that will be managed and used by
-//! the client to make other API requests.
-
-use super::views::DeviceAccessTokenGrant;
-use crate::ApiContext;
-use dropshot::Body;
-use dropshot::{HttpError, RequestContext};
-use http::{header, Response, StatusCode};
-use nexus_db_queries::db::model::DeviceAccessToken;
-use nexus_types::external_api::params;
-use serde::Serialize;
-
-// Token granting à la RFC 8628 (OAuth 2.0 Device Authorization Grant)
-
-/// OAuth 2.0 error responses use 400 (Bad Request) with specific `error`
-/// parameter values to indicate protocol errors (see RFC 6749 §5.2).
-/// This is different from Dropshot's error `message` parameter, so we
-/// need a custom response builder.
-pub(crate) fn build_oauth_response<T>(
-    status: StatusCode,
-    body: &T,
-) -> Result<Response<Body>, HttpError>
-where
-    T: ?Sized + Serialize,
-{
-    let body = serde_json::to_string(body)
-        .map_err(|e| HttpError::for_internal_error(e.to_string()))?;
-    Ok(Response::builder()
-        .status(status)
-        .header(header::CONTENT_TYPE, "application/json")
-        .body(body.into())?)
-}
-
-#[derive(Debug)]
-pub enum DeviceAccessTokenResponse {
-    Granted(DeviceAccessToken),
-    Pending,
-    #[allow(dead_code)]
-    Denied,
-}
-
-pub(crate) async fn device_access_token(
-    rqctx: RequestContext<ApiContext>,
-    params: params::DeviceAccessTokenRequest,
-) -> Result<Response<Body>, HttpError> {
-    let apictx = rqctx.context();
-    let nexus = &apictx.context.nexus;
-    let handler = async {
-        // RFC 8628 §3.4
-        if params.grant_type != "urn:ietf:params:oauth:grant-type:device_code" {
-            return build_oauth_response(
-                StatusCode::BAD_REQUEST,
-                &serde_json::json!({
-                    "error": "unsupported_grant_type"
-                }),
-            );
-        }
-
-        // RFC 8628 §3.5
-        let opctx = nexus.opctx_external_authn();
-        use DeviceAccessTokenResponse::*;
-        match nexus
-            .device_access_token_fetch(
-                &opctx,
-                params.client_id,
-                params.device_code,
-            )
-            .await
-        {
-            Ok(response) => match response {
-                Granted(token) => build_oauth_response(
-                    StatusCode::OK,
-                    &DeviceAccessTokenGrant::from(token),
-                ),
-                Pending => build_oauth_response(
-                    StatusCode::BAD_REQUEST,
-                    &serde_json::json!({
-                        "error": "authorization_pending"
-                    }),
-                ),
-                Denied => build_oauth_response(
-                    StatusCode::BAD_REQUEST,
-                    &serde_json::json!({
-                        "error": "access_denied"
-                    }),
-                ),
-            },
-            Err(error) => build_oauth_response(
-                StatusCode::BAD_REQUEST,
-                &serde_json::json!({
-                    "error": "invalid_request",
-                    "error_description": format!("{}", error),
-                }),
-            ),
-        }
-    };
-    apictx
-        .context
-        .external_latencies
-        .instrument_dropshot_handler(&rqctx, handler)
-        .await
-}
