+++ nexus/src/external_api/mod.rs
@@ -3,7 +3,6 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 pub mod console_api;
-pub mod device_auth;
 pub(crate) mod http_entrypoints;
 
 pub(crate) use nexus_types::external_api::params;
